Merged in master (pull request #55)
Master

* The Base64InputStreamTest class has been added. Test issues are:

- Unreachable code in the decode method. Code checks char array to be
multiplier of 4. But the calling method always send an array of lenght
4. Check may be removed.
* The Base64OutputStreamTest class has been added.
* CsvCallbackReaderTest class has been updated
* The JsonObjectMapTest class has been added. Test issues are:

- The "clear" method has a bug. ConcurrentModificationException is
thrown. I rewrote the method body.
- The "JsonObjectMap$entry" class has a bug. So I couldn't add tests for
the keySet, values, and entrySet methods. It doesn't implement the
Comparable interface. I think this class was never used in your code.
* The JsonVisitorTest class has been added. Test issues are:

- Unreachable code in the forEachElement methot. JsonArray contains null
values but cannot let them be read. As a result, the (o! = null) line is
unnecessary.

- Unreachable code in listContains metod. It uses the JSONArray.get()
method, which throws an exception when trying to reach null values. But
after this line, there is a redundant null check ( o != null ) As a
result, the null scenario is untestable.

- Unreachable code in the arrayToList metod. Because of the redundant
null check (t!= null), this line is untestable for the null scenario.
* The JsonEvalTest class has been updated.
* private constructor added
* CommentedJsonTokenerTest updated
* The CommentedJsonTokenerTest class has been added. Test issues are:

- Unreachable code can be found in some parts of the code. The
commenting functionality of a JSON file is not properly working. As a
result, I disabled the last tests. The programmer wants to use a single
slash as an escape character, but JSON can't parse these files.
* The JsonUtilTest class has been added. Test issues are:

- getIndexOfStringInArray method works with JSONArray objects. But it
works only String values. If JSONArray contains a numeric field, method
throws unexpected error. I think functionality not complete

Approved-by: developers @continual